refactor: extract events_ingest tests into separate file
